+++ WorkTimeTable.Tests/WorkerList.json
@@ -1,6 +1,6 @@
 ﻿[
   {
-    "BirthDate": "121212",
+    "BirthDate": "121111",
     "Brush": {
       "Color": "#FFDC143C",
       "Opacity": 1
@@ -10,21 +10,33 @@
     "Name": "AAA",
     "WorkTimes": [
       {
+        "StartWorkTime": "2024-01-01T00:00:00",
         "WorkerId": 0,
-        "WorkTimeType": 0,
-        "StartTime": "2024-01-01T00:00:00",
-        "WorkTimeSpan": "12:00:00"
+        "WorkTimeSpan": "08:00:00",
+        "WorkTimeType": 0
       },
       {
-        "WorkerId": 0,
-        "WorkTimeType": 0,
-        "StartTime": "2024-01-02T12:00:00",
-        "WorkTimeSpan": "12:00:00"
+        "StartWorkTime": "2024-01-01T20:00:00",
+        "WorkerId": 1,
+        "WorkTimeSpan": "09:00:00",
+        "WorkTimeType": 0
+      },
+      {
+        "StartWorkTime": "2024-01-03T13:00:00",
+        "WorkerId": 2,
+        "WorkTimeSpan": "12:00:00",
+        "WorkTimeType": 0
+      },
+      {
+        "StartWorkTime": "2024-01-05T23:00:00",
+        "WorkerId": 3,
+        "WorkTimeSpan": "23:00:00",
+        "WorkTimeType": 0
       }
     ]
   },
   {
-    "BirthDate": "121212",
+    "BirthDate": "111022",
     "Brush": {
       "Color": "#FF6495ED",
       "Opacity": 1
@@ -34,21 +46,33 @@
     "Name": "BBB",
     "WorkTimes": [
       {
-        "WorkerId": 1,
-        "WorkTimeType": 0,
-        "StartTime": "2024-01-01T12:00:00",
-        "WorkTimeSpan": "12:00:00"
+        "StartWorkTime": "2024-01-01T08:00:00",
+        "WorkerId": 0,
+        "WorkTimeSpan": "09:00:00",
+        "WorkTimeType": 0
       },
       {
+        "StartWorkTime": "2024-01-02T05:00:00",
         "WorkerId": 1,
-        "WorkTimeType": 0,
-        "StartTime": "2024-01-03T00:00:00",
-        "WorkTimeSpan": "12:00:00"
+        "WorkTimeSpan": "17:00:00",
+        "WorkTimeType": 0
+      },
+      {
+        "StartWorkTime": "2024-01-04T01:00:00",
+        "WorkerId": 2,
+        "WorkTimeSpan": "23:00:00",
+        "WorkTimeType": 0
+      },
+      {
+        "StartWorkTime": "2024-01-06T22:00:00",
+        "WorkerId": 3,
+        "WorkTimeSpan": "00:00:00",
+        "WorkTimeType": 0
       }
     ]
   },
   {
-    "BirthDate": "121212",
+    "BirthDate": "220923",
     "Brush": {
       "Color": "#FFFFD700",
       "Opacity": 1
@@ -58,16 +82,28 @@
     "Name": "CCC",
     "WorkTimes": [
       {
-        "WorkerId": 2,
-        "WorkTimeType": 0,
-        "StartTime": "2024-01-02T00:00:00",
-        "WorkTimeSpan": "12:00:00"
+        "StartWorkTime": "2024-01-01T17:00:00",
+        "WorkerId": 0,
+        "WorkTimeSpan": "03:00:00",
+        "WorkTimeType": 0
       },
       {
+        "StartWorkTime": "2024-01-02T22:00:00",
+        "WorkerId": 1,
+        "WorkTimeSpan": "15:00:00",
+        "WorkTimeType": 0
+      },
+      {
+        "StartWorkTime": "2024-01-05T00:00:00",
         "WorkerId": 2,
-        "WorkTimeType": 0,
-        "StartTime": "2024-01-03T12:00:00",
-        "WorkTimeSpan": "12:00:00"
+        "WorkTimeSpan": "23:00:00",
+        "WorkTimeType": 0
+      },
+      {
+        "StartWorkTime": "2024-01-06T22:00:00",
+        "WorkerId": 3,
+        "WorkTimeSpan": "22:00:00",
+        "WorkTimeType": 0
       }
     ]
   }
