+++ WorkTimeTable/WorkTimeTable.csproj
@@ -32,7 +32,7 @@
     <PackageReference Include="Serilog.Settings.Configuration" Version="8.0.0" />
     <PackageReference Include="Serilog.Sinks.Console" Version="5.0.1" />
     <PackageReference Include="Serilog.Sinks.File" Version="5.0.0" />
-    <PackageReference Include="SosoThemeLibrary" Version="1.0.16" />
+    <PackageReference Include="SosoThemeLibrary" Version="1.0.17" />
   </ItemGroup>
 
   <ItemGroup>
