+++ WorkTimeTable/WorkerList.json
@@ -1,6 +1,6 @@
 ﻿[
   {
-    "BirthDate": "121212",
+    "BirthDate": "121111",
     "Brush": {
       "Color": "#FFDC143C",
       "Opacity": 1
@@ -10,21 +10,33 @@
     "Name": "AAA",
     "WorkTimes": [
       {
-        "WorkerId": 1,
-        "WorkTimeType": 0,
         "StartWorkTime": "2024-01-01T00:00:00",
-        "WorkTimeSpan": "12:00:00"
+        "WorkerId": 1,
+        "WorkTimeSpan": "19:00:00",
+        "WorkTimeType": 0
       },
       {
+        "StartWorkTime": "2024-01-02T02:00:00",
         "WorkerId": 1,
-        "WorkTimeType": 0,
-        "StartWorkTime": "2024-01-02T12:00:00",
-        "WorkTimeSpan": "12:00:00"
+        "WorkTimeSpan": "09:00:00",
+        "WorkTimeType": 0
+      },
+      {
+        "StartWorkTime": "2024-01-04T02:00:00",
+        "WorkerId": 1,
+        "WorkTimeSpan": "04:00:00",
+        "WorkTimeType": 0
+      },
+      {
+        "StartWorkTime": "2024-01-05T04:00:00",
+        "WorkerId": 1,
+        "WorkTimeSpan": "19:00:00",
+        "WorkTimeType": 0
       }
     ]
   },
   {
-    "BirthDate": "121212",
+    "BirthDate": "111022",
     "Brush": {
       "Color": "#FF6495ED",
       "Opacity": 1
@@ -34,21 +46,33 @@
     "Name": "BBB",
     "WorkTimes": [
       {
+        "StartWorkTime": "2024-01-01T19:00:00",
         "WorkerId": 2,
-        "WorkTimeType": 0,
-        "StartWorkTime": "2024-01-01T12:00:00",
-        "WorkTimeSpan": "12:00:00"
+        "WorkTimeSpan": "04:00:00",
+        "WorkTimeType": 0
       },
       {
+        "StartWorkTime": "2024-01-02T11:00:00",
         "WorkerId": 2,
-        "WorkTimeType": 0,
-        "StartWorkTime": "2024-01-03T00:00:00",
-        "WorkTimeSpan": "12:00:00"
+        "WorkTimeSpan": "22:00:00",
+        "WorkTimeType": 0
+      },
+      {
+        "StartWorkTime": "2024-01-04T06:00:00",
+        "WorkerId": 2,
+        "WorkTimeSpan": "14:00:00",
+        "WorkTimeType": 0
+      },
+      {
+        "StartWorkTime": "2024-01-05T23:00:00",
+        "WorkerId": 2,
+        "WorkTimeSpan": "19:00:00",
+        "WorkTimeType": 0
       }
     ]
   },
   {
-    "BirthDate": "121212",
+    "BirthDate": "220923",
     "Brush": {
       "Color": "#FFFFD700",
       "Opacity": 1
@@ -58,16 +82,28 @@
     "Name": "CCC",
     "WorkTimes": [
       {
+        "StartWorkTime": "2024-01-01T23:00:00",
         "WorkerId": 3,
-        "WorkTimeType": 0,
-        "StartWorkTime": "2024-01-02T00:00:00",
-        "WorkTimeSpan": "12:00:00"
+        "WorkTimeSpan": "03:00:00",
+        "WorkTimeType": 0
       },
       {
+        "StartWorkTime": "2024-01-03T09:00:00",
         "WorkerId": 3,
-        "WorkTimeType": 0,
-        "StartWorkTime": "2024-01-03T12:00:00",
-        "WorkTimeSpan": "12:00:00"
+        "WorkTimeSpan": "17:00:00",
+        "WorkTimeType": 0
+      },
+      {
+        "StartWorkTime": "2024-01-04T20:00:00",
+        "WorkerId": 3,
+        "WorkTimeSpan": "08:00:00",
+        "WorkTimeType": 0
+      },
+      {
+        "StartWorkTime": "2024-01-06T18:00:00",
+        "WorkerId": 3,
+        "WorkTimeSpan": "07:00:00",
+        "WorkTimeType": 0
       }
     ]
   }
